Created new Directory for Report files
